Create eth0 & eth1 intf irrespective of vmi_if_count
This commit creates eth0/1 interfaces irrespective of
vmi_if_count value in the bios table.
Previously, it was implemented in such a way that
the interfaces will be created based on the if count
value in the bios table (if if_count is 1, the create eth0,
if if_count is 2, the  create eth0 & eth1) and hence, the
user was unable to create the eth1 interface at all (when
if_count is 1).

Also, itseems phyp doesnt take this bios property into account
and it has by default eth0 and eth1 interfaces.

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>
Change-Id: I241ed86d162dc501af740c3a6354fedb01d28a88