Donot reset vmi ip properties on origin change
When the origin is changed to static/dhcp, the corresponding
attribute in the biostable (vmi_if<0/1>_ipv4_method) will be updated
with the same and this takes care of updating the new values
that are sent down from vmi in the hypervisor netword dbus object.
The call to reset ip properties is causing additional updation
of the properties with 0s, which is not necessary here.

Tested By:

* Set dhcp to true on an ethernet interface:
PATCH -d '{"DHCPv4": {"DHCPEnabled" :true}}' https://${bmc}/redfish/v1/Systems/hypervisor/EthernetInterfaces/eth1

* Checked the biostable ip values and checked if the hypervisor ip
dbus object is in sync with bios table.

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>