Allow static config of same ip & enhance logs
There were scenarios where the host is not in
the required state to communicate ip configuration
details to the host. In that case, the dbus contains
the updated values but the enabled property will be
false and the client might try to reconfigure the
same ip, which will not be allowed in the hypervisor
app.
This PR will allow the client to configure the same ip
again, thus the hypervisor app will update the bios table
which will be communicated to the host when the host is
up.

This PR also enhances the log by printing the values of
the ip properties that were updated.

Tested By:

* Configure a static ip
* Reconfigure ip once again and see if the enabled  property
  on dbus is set
* Checked if the values are printed in the logs

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>
Change-Id: Id56d71a998d38b3e2d8117b0519d537163c7f8c1