BMC factory reset: Remove linklocal IP from eth1 (#39)
When linklocal is set on both interfaces after a factory reset, the BMC
comes up with two default routes for eth0 and eth1 on same subnet. This
causes confusions in the initial routing table, when the physical cable
connection is made from a local laptop. The connection drops are random,
sometimes eth0 works and sometimes eth1 works.

This commit is a workaround to disable linklocal assignment to eth1.
This will get only eth0 interface up with a linklocal address after a
factory reset. The actual fix will be worked with systemd upstream
community.

Tested by:
1. Factory reset, and see eth1 is not assigned with any linklocal IP
   Only eth0 interface will get the linklocal assigned

Signed-off-by: Sunitha Harish <sunithaharish04@gmail.com>
Change-Id: I4997842d49c525f1609c63bb9ebeb7df330d2a30