Fix synchronization issue in hypervisor app
This issue is caused by a timing problem during the synchronization
between the bios manager and hypervisor network manager, when there
is an ip configuration. The dbus notification from the bios manager
will be received at the hypervisor manager, while the hyp-mgr is still
winding up the static settings.

We have ipv6 settings as well now which may have caused this fraction
of delays.

Tested By:

* Static IP configuration on ethernet interface
* Switching to dhcp config on the same eth interface

Synchronization happens as expected.

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>