Fixed unwanted dhcp6 enabling on adding static ipv4
This scenario is related to one of the cases where we avoid coexistence of
DHCPv4 and static IPv4 address.

Currently, when DHCPv4 is enabled & DHCPv6 is disabled, on adding a static
v4 address DHCPv6 attribute is getting enabled

This commit will fix enabling of dhcv6 for the above scenario.

Tested by:
HCPv4 is enabled & DHCPv6 is disabled, by adding static IPv4 address
DHCPv6 attribute is not getting enabled.