inventory_mac: Fix Update MAC for eth1
Whenever the network daemon starts, and if the inventory manager has
already started, then networkd should read the MAC address from the
inventory and update the dbus object and configuration file.

Currently, this is being done for just eth0 interface and is exiting
before updating the MAC on eth1.

This commit fixes the above issue.

Tested-By:
* Restarted network manager service; checked MAC on both
inventory and network dbus & config files.
* Checked for both reboot and factory reset usecases.

Change-Id: I49d94f22d914f74c23999a5fe175bfc4d97759a9
Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>