Fix for Network Manager Crash (#129)
RTNet Link Packet is received with a buffer size of 4096 bytes, the
pointer to the buffer is not moved back to the start address of buffer
when the pointer to the buffer is used to receive the next packet in
queue causing the pointer access of array out of bound. The fix for this
in this patch re-initializes the pointer to start address of buffer
before the next packet is received. No other checking is required to
validate the case of array out of bound as the NLMSG preprocessor
macros is taking care of checking the received and remaining length of
data in the packet and send for processing.

Tested:
Flashed the BMC image on available platforms in which the crash was
reported and regression of sending the Net Link Packet continuously.
No core file/crash was observed during the time of regression and
reboot cycle tests.


Change-Id: Ibcfbb03dcce694d4116163be3db25f0597cda1eb

Signed-off-by: sureshvijayv1 <suresh.vijayakumar@intel.com>
Co-authored-by: sureshvijayv1 <suresh.vijayakumar@intel.com>