Handle IPv4 address for IPv6 Static Default Gateway
Error is not thrown when an IPV4 address is provided.
This commit addresses the issue by explicitly handling such cases
and returning an InvalidArgument error when a IPV4address
is configured for IPV6 Static Default Gateway field.

Tested By:
Verified the test case and ensured proper
invalid argument error is thrown.

Change-Id: I3925a40840d20c3d44a44fb6e2ef93672a4ef69e
Signed-off-by: kokilav <kokilavaradhan@gmail.com>