Handle IPv4 address for IPv6 Static Default Gateway
Error is not thrown when an IPV4 address is provided.
This commit addresses the issue by explicitly handling such cases
and returning an InvalidArgument error when a IPV4address
is configured for IPV6 Static Default Gateway field.

Tested By:
Verified the test case and ensured proper
invalid argument error is thrown.

Change-Id: I4074f939ffd23c533fac1661f2243322d0cac62a
Signed-off-by: kokilav <kokilavaradhan@gmail.com>